OV-396: * avatar-video controller
